# perf(index): use the `compiler`'s cached `fs` for stats (`this.fs.stat`)