@gusvargas 2018-10-23
If anyone is interested in trying this change out I created a fork of an open source app here: https://github.com/gusvargas/hackernews-react-graphql

Steps to reproduce:
1. `yarn && yarn dev`
2. Navigate to `localhost:3000` (I think there is some magic in nextjs that makes it so that only "active" modules are compiled)
3. Make a trivial change to "Main.js" and observe the terminal output. For me, compile time was floating slightly above and below 100ms.
4. Uncomment the webpack config change in `next.config.js` to enable the cache-loader
5. Restart the dev server and repeat step 3. Now I see compile times around 600ms-700ms.

Linking against the changes in PR brings the compile times back down to what is observed in step 3.