@gusvargas 2018-10-26
@michael-ciniawsky Any chance we can get the 1.2.3 release soon? The incremental build performance issue was the only thing holding us back from using the `cache-loader` – we'd love to start taking advantage of it ASAP.