@michael-ciniawsky 2018-10-30
Sry for the delay, released in [`v1.2.3`](https://github.com/webpack-contrib/cache-loader/releases/tag/v1.2.3) 🎉